apply dialog library to reviews
