+++ .prettierignore
@@ -1,2 +1,1 @@
 .next
-src/lib/*
