+++ src/modules/App/context/index.tsx
@@ -6,10 +6,12 @@ import LocalizationProvider from '@mui/lab/LocalizationProvider'
 import { ThemeProvider, useMediaQuery } from '@mui/material'
 
 import React from 'react'
+import { Toaster, ToastOptions } from 'react-hot-toast'
 
 import { AnalyticsProvider } from '@/common/context/Analytics'
 import { useAnalytics } from '@/common/context/Analytics/hooks/useAnalytics'
 import { darkTheme, lightTheme } from '@/configs/theme'
+import { Dialog } from '@/lib/dialog'
 import { ShoppingCartModalContextProvider } from '@/modules/App/context/ShoppingCartModal'
 import { SnackbarContextProvider } from '@/modules/App/context/Snackbar'
 import { client } from '@/services/apollo'
@@ -32,6 +34,16 @@ export function AppProvider({ children, forceDark, emotionCache }: AppProviderPr
 
   const { addEvent } = useAnalytics()
 
+  const toastOptions: ToastOptions = {
+    style: {
+      top: 20,
+      fontSize: '0.9rem',
+      fontFamily: 'prompt',
+      borderRadius: 4,
+      padding: 12,
+    },
+  }
+
   return (
     <ApolloProvider client={client}>
       <CacheProvider value={emotionCache}>
@@ -39,6 +51,8 @@ export function AppProvider({ children, forceDark, emotionCache }: AppProviderPr
           <AnalyticsProvider value={{ addEvent }}>
             <SnackbarContextProvider>
               <ThemeProvider theme={prefersDarkMode || forceDark ? darkTheme : lightTheme}>
+                <Dialog />
+                <Toaster position="top-center" toastOptions={toastOptions} gutter={16} />
                 <ShoppingCartModalContextProvider>{children}</ShoppingCartModalContextProvider>
               </ThemeProvider>
             </SnackbarContextProvider>
